Move package management helpers out of command.ts into dedicated files
