+++ packages/create-cloudflare/src/cli.ts
@@ -7,13 +7,12 @@ import { dim } from "@cloudflare/cli/colors";
 import { isInteractive } from "@cloudflare/cli/interactive";
 import { parseArgs } from "helpers/args";
 import { C3_DEFAULTS, isUpdateAvailable } from "helpers/cli";
+import { installWrangler, runCommand } from "helpers/command";
 import {
-	installWrangler,
-	npmInstall,
+	detectPackageManager,
 	rectifyPmMismatch,
-	runCommand,
-} from "helpers/command";
-import { detectPackageManager } from "helpers/packages";
+} from "helpers/packageManagers";
+import { npmInstall } from "helpers/packages";
 import { version } from "../package.json";
 import {
 	gitCommit,
