+++ packages/create-cloudflare/templates/pre-existing/c3.ts
@@ -4,7 +4,7 @@ import { join } from "path";
 import { processArgument } from "@cloudflare/cli/args";
 import { brandColor, dim } from "@cloudflare/cli/colors";
 import { runCommand } from "helpers/command";
-import { detectPackageManager } from "helpers/packages";
+import { detectPackageManager } from "helpers/packageManagers";
 import { chooseAccount } from "../../src/common";
 import type { C3Context } from "types";
