+++ packages/create-cloudflare/templates/remix/c3.ts
@@ -3,7 +3,7 @@ import { brandColor, dim } from "@cloudflare/cli/colors";
 import { spinner } from "@cloudflare/cli/interactive";
 import { transformFile } from "helpers/codemod";
 import { runFrameworkGenerator } from "helpers/command.js";
-import { detectPackageManager } from "helpers/packages";
+import { detectPackageManager } from "helpers/packageManagers";
 import type { TemplateConfig } from "../../src/templates";
 import type { C3Context } from "types";
