+++ packages/create-cloudflare/templates/vue/c3.ts
@@ -1,5 +1,5 @@
 import { runFrameworkGenerator } from "helpers/command";
-import { detectPackageManager } from "helpers/packages";
+import { detectPackageManager } from "helpers/packageManagers";
 import type { TemplateConfig } from "../../src/templates";
 import type { C3Context } from "types";
