+++ packages/create-cloudflare/vitest.d.ts
@@ -6,6 +6,7 @@ import 'vitest'
 
 interface CustomMatchers<R = unknown> {
 	toExist(): R;
+	toHaveBeenCalledFirstWith(args: string[]): R;
 }
 
 declare module 'vitest' {
